Rename utils.py to mixins.py. More precise.
https://github.com/python-discord/site/issues/305
