+++ pydis_site/apps/api/models/__init__.py
@@ -15,4 +15,4 @@ from .bot import (
     User
 )
 from .log_entry import LogEntry
-from .utils import ModelReprMixin
+from .mixins import ModelReprMixin
