+++ pydis_site/apps/api/models/bot/documentation_link.py
@@ -1,6 +1,6 @@
 from django.db import models
 
-from pydis_site.apps.api.models.utils import ModelReprMixin
+from pydis_site.apps.api.models.mixins import ModelReprMixin
 
 
 class DocumentationLink(ModelReprMixin, models.Model):
