+++ pydis_site/apps/api/models/bot/message_deletion_context.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 from pydis_site.apps.api.models.bot.user import User
-from pydis_site.apps.api.models.utils import ModelReprMixin
+from pydis_site.apps.api.models.mixins import ModelReprMixin
 
 
 class MessageDeletionContext(ModelReprMixin, models.Model):
