+++ pydis_site/apps/api/models/bot/off_topic_channel_name.py
@@ -1,7 +1,7 @@
 from django.core.validators import RegexValidator
 from django.db import models
 
-from pydis_site.apps.api.models.utils import ModelReprMixin
+from pydis_site.apps.api.models.mixins import ModelReprMixin
 
 
 class OffTopicChannelName(ModelReprMixin, models.Model):
