+++ pydis_site/apps/api/models/bot/reminder.py
@@ -2,7 +2,7 @@ from django.core.validators import MinValueValidator
 from django.db import models
 
 from pydis_site.apps.api.models.bot.user import User
-from pydis_site.apps.api.models.utils import ModelReprMixin
+from pydis_site.apps.api.models.mixins import ModelReprMixin
 
 
 class Reminder(ModelReprMixin, models.Model):
